+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.3.0] - 2025-03-26
+
+### 📚 Documentation
+
+- préparation pour la version 0.2.2
+- préparation pour la version 0.2.1
+- préparation pour la version 0.2.0
+
+### 🔄 Autres changements
+
+- push2Git & ignore credentials
+- Add files via upload
+- Initial commit
+
+
+
 ## [0.2.2] - 2025-03-26
 
 ### 📚 Documentation